# Avoid current_exe() and path canonicalisation when mapping.