--- a/ykutil/Cargo.toml
+++ b/ykutil/Cargo.toml
@@ -10,3 +10,6 @@ libc = "0.2.117"
 memmap2 = "0.5.2"
 object = "0.28.3"
 phdrs = { git = "https://github.com/softdevteam/phdrs" }
+
+[build-dependencies]
+cc = "1.0.72"