--- a/ykutil/build.rs
+++ b/ykutil/build.rs
@@ -1,0 +1,5 @@
+fn main() {
+    let mut comp = cc::Build::new();
+    comp.file("src/find_main.c");
+    comp.compile("find_main");
+}