--- a/ykutil/src/addr.rs
+++ b/ykutil/src/addr.rs
@@ -1,32 +1,15 @@
 //! Address utilities.
 
+use crate::obj::SELF_BIN_PATH;
 use libc::{c_void, dladdr, Dl_info};
 use phdrs::objects;
 use std::mem::MaybeUninit;
 use std::{
     convert::TryFrom,
-    env,
     ffi::CStr,
-    fs,
     path::{Path, PathBuf},
     ptr::{null, null_mut},
 };
-
-// Make the path `p` a canonical absolute path.
-fn canonicalise_path(p: &CStr) -> PathBuf {
-    // On linux the empty object path means the "main binary".
-    #[cfg(target_os = "linux")]
-    if p.to_str().unwrap() == "" {
-        return PathBuf::new();
-    }
-    let p_path = Path::new(p.to_str().unwrap());
-    if p.to_str().unwrap() == "linux-vdso.so.1" {
-        // The VDSO isn't a real file that can be canonicalised.
-        p_path.to_owned()
-    } else {
-        fs::canonicalize(p_path).unwrap()
-    }
-}
 
 /// Given a virtual address, returns a pair indicating the object in which the address originated
 /// and the byte offset.
@@ -41,18 +24,15 @@ pub fn code_vaddr_to_off(vaddr: usize) -> Option<(PathBuf, u64)> {
     if unsafe { dladdr(vaddr as *const c_void, &mut info as *mut Dl_info) } == 0 {
         return None;
     }
-    let containing_obj = canonicalise_path(unsafe { CStr::from_ptr(info.dli_fname) });
+    let containing_obj = PathBuf::from(unsafe { CStr::from_ptr(info.dli_fname) }.to_str().unwrap());
 
     // Find the corresponding byte offset of the virtual address in the object.
     for obj in &objects() {
         let obj_name = obj.name();
-        let obj_name: PathBuf = if unsafe { *obj_name.as_ptr() } == 0 {
-            // On some systems, the empty string indicates the main binary.
-            let exe = env::current_exe().unwrap();
-            debug_assert_eq!(&fs::canonicalize(&exe).unwrap(), &exe);
-            exe
+        let obj_name: &Path = if unsafe { *obj_name.as_ptr() } == 0 {
+            SELF_BIN_PATH.as_path()
         } else {
-            canonicalise_path(obj_name)
+            Path::new(obj_name.to_str().unwrap())
         };
         if obj_name != containing_obj {
             continue;
@@ -74,7 +54,7 @@ pub fn off_to_vaddr_main_obj(off: u64) -> Option<usize> {
 /// Find the virtual address of the offset `off` in the object `containing_obj`.
 pub fn off_to_vaddr(containing_obj: &Path, off: u64) -> Option<usize> {
     for obj in &objects() {
-        if canonicalise_path(obj.name()) != containing_obj {
+        if Path::new(obj.name().to_str().unwrap()) != containing_obj {
             continue;
         }
         return Some(usize::try_from(off + obj.addr()).unwrap());