This basically works by `replacing std::env::current_exe()` by querying the linker for the object containing `main()` and using that path exclusively.

This has the knock-on effect that we no longer have one mechanism giving us a non-canonical path (`dladdr(3)`, when asking where a virtual address came from), and one mechanism giving us (possibly) an absolute path (`std::env::current_exe()`). Therefore we no longer need to canonicalise paths, which is a win in itself because canonicalisation hits the filesystem, slowing us down.

Running numwarp.bf on the input `1234`:

BEFORE:
```
1: sh -c "echo \"1234\" |  ./bf_simple_yk lang_tests/numwarp.bf"
            Mean        Std.Dev.    Min         Median      Max
real        1.638       0.011       1.623       1.642       1.654
user        1.241       0.014       1.218       1.244       1.256
sys         0.348       0.014       0.331       0.344       0.369
```

AFTER:
```
1: sh -c "echo \"1234\" |  ./bf_simple_yk lang_tests/numwarp.bf"
            Mean        Std.Dev.    Min         Median      Max
real        1.479       0.007       1.471       1.474       1.489
user        1.194       0.013       1.172       1.197       1.208
sys         0.235       0.016       0.210       0.233       0.259
```

So about 10% quicker, which is close enough to the ~7% cost of canonicalisation that I measured using perf.

(this is with Lukas' experimental PT cache and guard compression diffs applied, as without those, the cost of canonicalisation is swamped by other inefficiencies).