@vext01 2022-09-14
I haven't spent much time tidying this. I suspect I can probably refactor some of the surrounding code given more time.