@vext01 2022-09-14
> Any reason this is a draft PR? It looks mergeable to me!

Well! I thought you would object to the hackiness of it all.

Also:

> I haven't spent much time tidying this. I suspect I can probably refactor some of the surrounding code given more time.

We can merge it if you like, and I can look to do more tidying tomorrow in another PR.