@vext01 2022-09-16
splat.