@vext01 2022-09-16
> error: entry symbol `main` declared multiple times

well, that's a new one on me.