@vext01 2022-09-16
I was only able to reproduce this after upgrading Rust!

Then I boiled it down to this (in a lib crate):

```
use libc::{c_char, c_int};

extern "C" {
    fn main(argc: c_int, argv: *const *const c_char) -> c_int;
}

pub fn f() {
    dbg!(main as *const u8);
}

#[cfg(test)]
mod tests {
    use super::f;

    #[test]
    fn it_works() {
        f();
    }
}
```

Running `cargo test` will make it explode.

```
...
   Compiling ppp v0.1.0 (/tmp/ppp)
error: entry symbol `main` declared multiple times
  |
  = help: did you use `#[no_mangle]` on `fn main`? Use `#[start]` instead

```

My guess: some kind of clash with the test binary main, or something like that?