@vext01 2022-09-20
This tuned out to be a Rust bug.

https://github.com/rust-lang/rust/issues/101906

I've worked around it with C code for now.

OK to squash?