@vext01 2022-09-21
splat.